+++ ZDOInfo/ContainerZDOInfoProvider.cs
@@ -9,8 +9,8 @@ namespace ValheimRcon.ZDOInfo
         public override void AppendInfo(ZDO zdo, StringBuilder stringBuilder, bool detailed)
         {
             TempInventory.RemoveAll();
-            var itemsData = zdo.GetString(ZDOVars.s_items);
-            if (!string.IsNullOrEmpty(itemsData))
+            var itemsData = zdo.GetByteArray(ZDOVars.s_items, null);
+            if (itemsData != null)
             {
                 var package = new ZPackage(itemsData);
                 TempInventory.Load(package);